requirements.txt
tqdm
timm==0.6.5
ftfy
numpy
regex
pyyaml
scikit-learn
torch==1.13.0
